+++ src/Extension/CommonMark/Parser/Inline/AutolinkParser.php
@@ -25,30 +25,30 @@ use League\CommonMark\Util\UrlEncoder;
 final class AutolinkParser implements InlineParserInterface
 {
     private const EMAIL_REGEX      = '<([a-zA-Z0-9.!#$%&\'*+\\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*)>';
-    private const OTHER_LINK_REGEX = '<[A-Za-z][A-Za-z0-9.+-]{1,31}:[^<>\x00-\x20]*>';
+    private const OTHER_LINK_REGEX = '<([A-Za-z][A-Za-z0-9.+-]{1,31}:[^<>\x00-\x20]*)>';
 
     public function getMatchDefinition(): InlineParserMatch
     {
         return InlineParserMatch::regex(self::EMAIL_REGEX . '|' . self::OTHER_LINK_REGEX);
     }
 
-    public function parse(string $match, InlineParserContext $inlineContext): bool
+    public function parse(InlineParserContext $inlineContext): bool
     {
-        $cursor = $inlineContext->getCursor();
-        if ($m = $cursor->match('/^' . self::EMAIL_REGEX . '/')) {
-            $email = \substr($m, 1, -1);
-            $inlineContext->getContainer()->appendChild(new Link('mailto:' . UrlEncoder::unescapeAndEncode($email), $email));
+        $inlineContext->getCursor()->advanceBy($inlineContext->getFullMatchLength());
+        $matches = $inlineContext->getMatches();
+
+        if ($matches[1] !== '') {
+            $inlineContext->getContainer()->appendChild(new Link('mailto:' . UrlEncoder::unescapeAndEncode($matches[1]), $matches[1]));
 
             return true;
         }
 
-        if ($m = $cursor->match('/^' . self::OTHER_LINK_REGEX . '/')) {
-            $dest = \substr($m, 1, -1);
-            $inlineContext->getContainer()->appendChild(new Link(UrlEncoder::unescapeAndEncode($dest), $dest));
+        if ($matches[2] !== '') {
+            $inlineContext->getContainer()->appendChild(new Link(UrlEncoder::unescapeAndEncode($matches[2]), $matches[2]));
 
             return true;
         }
 
-        return false;
+        return false; // This should never happen
     }
 }