+++ src/Extension/CommonMark/Parser/Inline/HtmlInlineParser.php
@@ -29,10 +29,12 @@ final class HtmlInlineParser implements InlineParserInterface
         return InlineParserMatch::regex(RegexHelper::PARTIAL_HTMLTAG);
     }
 
-    public function parse(string $match, InlineParserContext $inlineContext): bool
+    public function parse(InlineParserContext $inlineContext): bool
     {
-        $inlineContext->getCursor()->advanceBy(\mb_strlen($match));
-        $inlineContext->getContainer()->appendChild(new HtmlInline($match));
+        $inline = $inlineContext->getFullMatch();
+
+        $inlineContext->getCursor()->advanceBy($inlineContext->getFullMatchLength());
+        $inlineContext->getContainer()->appendChild(new HtmlInline($inline));
 
         return true;
     }